+++ src/common/exec_util.h
@@ -31,8 +31,8 @@ struct block
 {
     int nnz;
     int roffset, coffset;
-    //unsigned short int *rloc, *cloc;
-    int *rloc , *cloc ; 
+    unsigned short int *rloc, *cloc;
+    //int *rloc , *cloc ; 
     double *val;
 };
